refactor: tr_sessionClose() now takes fractions of seconds for deadline
refactor: make tr_web::startShutdown() deadline dependent on the tr_sessionClose() deadline arg

+++ libtransmission/session.h
@@ -1281,7 +1281,7 @@ private:
     friend uint16_t tr_sessionGetPeerPort(tr_session const* session);
     friend uint16_t tr_sessionGetRPCPort(tr_session const* session);
     friend uint16_t tr_sessionSetPeerPortRandom(tr_session* session);
-    friend void tr_sessionClose(tr_session* session, size_t timeout_secs);
+    friend void tr_sessionClose(tr_session* session, double timeout_secs);
     friend tr_variant tr_sessionGetSettings(tr_session const* s);
     friend void tr_sessionLimitSpeed(tr_session* session, tr_direction dir, bool limited);
     friend void tr_sessionReloadBlocklists(tr_session* session);
